Restructure packaging for bacs
COAND-497
